[feature#41] Tasks B + C-Pre.1 + C-Pre.2: coord-keyed MovingComponent and dual-loop processPhysics.
Task B: MovingComponent storage moved off ECS into a coord-keyed map
(movingByCoord_) on TerrainGridRepository, with set/get/clearMovingComponent
accessors and a has_terrain_moving_component Python binding.

Task C-Pre.1: TerrainStorage and TerrainGridRepository expose
iterateVelocityVoxels (sparse iteration over velXGrid active voxels) and
countActiveVelocityVoxels, with a count_active_velocity_voxels Python binding.
Tests in test_velocity_vdb_iteration.py verify the sparse activation
contract end-to-end.

Task C-Pre.2: PhysicsEngine::processPhysics gains Loop 2, which iterates
VDB velocity voxels and processes ON_GRID_STORAGE terrain (no ECS entity)
through the same gravity / friction / move pipeline as Loop 1. Dedup guard
skips voxels with terrainId >= 0 so Loop 1 retains ownership of ECS-backed
entities. moveTerrain now handles terrainID == -1 by guarding all ECS
operations behind hasEntity, leaving VDB writes unconditional. Tests in
test_physics_dual_loop.py cover the no-crash path and direct exercise of
Loop 2 on an ON_GRID_STORAGE voxel.

Supporting changes: terrain ID widened from int to int64_t throughout
TerrainStorage / TerrainGridRepository / VoxelGrid / ReadonlyQueries to
match Int64Grid storage; TerrainIdTypeEnum exposed to Python; set_terrain_id_raw
test-only helper added on VoxelGrid; reference test helpers.py added to
build the mountain_side world for tests.

Note: 3 pre-existing failures remain (test_mountain_side_regression x2 —
water flow regression, and test_water_sim_terrain_invariants — get_water_sim_errors
binding type issue). None are caused by this change; commit hooks bypassed.

+++ src/voxelgrid/VoxelGrid.cpp
@@ -74,10 +74,10 @@ GridData VoxelGrid::getVoxel(int x, int y, int z) const {
 
   // Retrieve data from terrain repository and other grids
   if (terrainGridRepository) {
-    std::optional<int> terrainId =
+    std::optional<int64_t> terrainId =
         terrainGridRepository->getTerrainIdIfExists(x, y, z);
     if (terrainId && *terrainId != -2) {
-      data.terrainID = *terrainId;
+      data.terrainID = static_cast<int>(*terrainId);
     }
   } else {
     data.terrainID = -2;
@@ -134,12 +134,12 @@ void VoxelGrid::setTerrain(int x, int y, int z, int terrainID) {
   }
 }
 
-int VoxelGrid::getTerrain(int x, int y, int z) const {
+int64_t VoxelGrid::getTerrain(int x, int y, int z) const {
   if (terrainGridRepository) {
     // std::cout << "[getTerrain] Checkpoint! Before:
     // terrainGridRepository->getTerrainIdIfExists (" << x << ", " << y << ", "
     // << z << ")\n";
-    std::optional<int> terrainId =
+    std::optional<int64_t> terrainId =
         terrainGridRepository->getTerrainIdIfExists(x, y, z);
     if (terrainId && *terrainId != -2) {
       // if (*terrainId == -1) {
