+++ src/voxelgrid/VoxelGrid.hpp
@@ -58,7 +58,7 @@ public:
                                      const MatterContainer &component) const;
 
   void setTerrain(int x, int y, int z, int terrainID);
-  int getTerrain(int x, int y, int z) const;
+  int64_t getTerrain(int x, int y, int z) const;
   int createEnttForTerrain(int x, int y, int z) const;
   // Forward to TerrainGridRepository::deleteTerrain. If `takeLock` is false,
   // the repository will assume the caller already holds the terrain grid lock.
